Split hosted auth config module
